+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - The AccountCapabilities class
 - The MailCapabilities class
 - Refactored Account capabilities to match the spec. #13
+- Account.hasMail
+- Account.hasCalendars
+- Account.hasContacts
 
 ## [0.0.13] - 2016-02-24
 ### Fixed